+++ visualization/app/codeCharta/state/store/dynamicSettings/colorRange/resetColorRange.effect.ts
@@ -1,16 +1,17 @@
 import { Injectable } from "@angular/core"
-import { Store, State } from "@ngrx/store"
+import { Store } from "@ngrx/store"
 import { Actions, createEffect, ofType } from "@ngrx/effects"
-import { map, skip, switchMap, take, tap } from "rxjs"
+import { map, skip, switchMap, take } from "rxjs"
 import { selectedColorMetricDataSelector } from "../../../selectors/accumulatedData/metricData/selectedColorMetricData.selector"
 import { calculateInitialColorRange } from "./calculateInitialColorRange"
 import { setColorRange } from "./colorRange.actions"
 import { fileActions } from "../../files/files.actions"
 import { CcState } from "../../../../codeCharta.model"
+import { setColorMetric } from "../colorMetric/colorMetric.actions"
 
 @Injectable()
 export class ResetColorRangeEffect {
-	constructor(private actions$: Actions, private store: Store<CcState>, private state: State<CcState>) {}
+	constructor(private actions$: Actions, private store: Store<CcState>) {}
 
 	resetColorRange$ = createEffect(() =>
 		this.actions$.pipe(
@@ -20,15 +21,11 @@ export class ResetColorRangeEffect {
 		)
 	)
 
-	resetColorRangeOnColorMetricChange$ = createEffect(
-		() =>
-			this.actions$.pipe(
-				ofType("SET_COLOR_METRIC"),
-				tap(() => {
-					const selectedColorMetricData = selectedColorMetricDataSelector(this.state.getValue())
-					this.store.dispatch(setColorRange({ value: calculateInitialColorRange(selectedColorMetricData) }))
-				})
-			),
-		{ dispatch: false }
+	resetColorRangeOnColorMetricChange$ = createEffect(() =>
+		this.actions$.pipe(
+			ofType(setColorMetric),
+			switchMap(() => this.store.select(selectedColorMetricDataSelector).pipe(take(1))),
+			map(selectedColorMetricData => setColorRange({ value: calculateInitialColorRange(selectedColorMetricData) }))
+		)
 	)
 }